requirements.txt
python==3.8.6
numpy==1.19.0
scipy==1.6.0
networkx==2.5
matplotlib==3.3.3
joblib==0.17.0
pyyaml==5.3.1
tqdm==4.54.1